Change SectionContactList:  rename, add show modal logic
