+++ src/components/Main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import SectionAddButton from './SectionAddButton';
-import ContactList from './ContactList';
+import SectionContactList from './SectionContactList';
 
 import s from './Main.module.scss';
 
@@ -9,9 +9,7 @@ const Main = () => (
   <main className={s.main}>
     <h1 className={s.title}>Contacts</h1>
     <SectionAddButton />
-    <section className={s.contacts}>
-      <ContactList />
-    </section>
+    <SectionContactList />
   </main>
 );
