+++ src/components/Main.module.scss
@@ -11,22 +11,3 @@
   margin-top: 100px;
   color: $orange200;
 }
-
-.addSection {
-  position: relative;
-  min-width: 360px;
-  width: 84%;
-  text-align: center;
-  @media (min-width: $tablet-lg-width) {
-    width: 60%;
-  }
-
-  @media (min-width: $desktop-md-width) {
-    width: 900px;
-  }
-}
-
-.contacts {
-  width: 100%;
-  min-width: 320px;
-}
