+++ src/components/SectionContactList.module.scss
@@ -6,3 +6,8 @@
   width: 88%;
   list-style-type: none;
 }
+
+.contacts {
+  width: 100%;
+  min-width: 320px;
+}
